fix broken link in docs
Prior to this commit, the documenation hosted at hexdocs.pm would render
the wrong link for the heroicons website because the markup was lacking
the protocol scheme, which would cause the link to be interpreted as a
relative link to https://hexdocs.pm/heroicons/heroicons.com rather than
one out to https://heroicons.com. This commit fixes both the generated
`lib/heroicons.ex` and the source `assets/heroicons.exs`.